Add 'ResourceGovernance' note for other settings
The additional properties noted in the section beginning on line 69
are not in the ThreadPool property node, unlike the SchedulingBehavior
property. The only way to discover the correct property node is to
open up the msmdsrv.ini config file for SSAS.

Adding this note and XMLA snippet will make it easier for others to
find and update these settings.