Update 'RowsetSerializationLimit with warning on 0
Setting the property value of OLAP\Query\RowsetSerializationLimit to 0
will cause all queries returning rowsets to fail. This includes the
XMLA queries to change exactly this setting, leaving the SSAS instance
in a nigh-unusable state.

The only way that it seems possible to revert such a change (to 0) is
to manually edit the config file for SSAS and restart the service.